requirements.txt
img2vec-pytorch
pillow
scikit-learn
joblib